Quick adjustment to SkillsGrid grid styles
I wanted to make it so that the top and bottom rows remained open (top row has no border, bottom row has no border) and the current implementation made that difficult to accomplish considering that rows will not always be even due to responsiveness and data returned. The quick solution was to switch the border edges from using a bottom edge to a top edge so that we could select specifically the first row and unset that border, which is a lot simpler than calculating how many items ended up in the final row.