Separating Carousel Feature into two views for better mobile exp and less cluttered screen
This is a rework of the Carousel feature after reviewing how it would scale to mobile and also having some second thoughts about how much I was trying to pack into a single view. While it can be nice to simply click through different projects and have immediate access to their carousels without ever leaving the view, it also just means there is a lot on the screen at once and it can be overwhelming and feel less focused. Furthermore, it is not going to be easy to fit all of that on a single mobile view and rather than have two different paradigms for different viewports I would like to create the best user experience that translates effectively from large to small scale.

In addition to the Carousel updates, I have also made some adjustments in a handful of other places to refine visual design, fix some issues, etc.