Removing some code that is no longer needed
This code was from early tests for handling mobile menu/navigation. I was going to have a separate menu button for mobile to review links, but then I realized that the paradigm this whole app is using is that the main screen IS the menu, so I replaced the X to close views with a menu icon instead. This makes sense and works well in both mobile and desktop environments