Reworks max card view title/head area
I have been going back and forth on whether or not to have a header that content scrolls behind or that pushes content down - aka whether the scrollbar should be full view height or start directly under the header.
After trying a few things, I have settled on having full scrolling. The goal for this application is to avoid scrolling on most desktop and laptop devices by designing content to be minimal and fit nicely within a single viewport. However, for cases where scrolling is necessary, a backdrop blur on the logo and title elements will make the elements behind the header feel less like they are conflicting/overlapping in that space. The end result is quite nice.