Hotfix - prevents state update conflict between ProjectBrowser and ProjectCarousel
In a recent update, I had made an update to the project browser so that when we use the go back button it sets the selected project to null. This seemed like the right thing to do at the time because technically when you close the ProjectCarousel to return to the Browse Selection screen you technically are no longer looking at a specific project. However, this was not needed because the ProjectBrowser is built in such a way that only one view can ever be rendered at a time. So this caused an issue with the ProjectCarousel state as it would try to update when the ProjectId changed but simultaneously its parent component was removing it from render. To resolve this, I did the following:
- Removed the explicit setting of null. Since ProjectBrowser is removing the component from the render tree, its props are no longer relevant.
- Added a null check to ProjectCarousel just as a failsafe. It will always return null if ProjectId is null or undefined.

Also:
- Fine tuned colors some more