ProjectBrowser and ProjectCarousel Bug fixed - for real this time!
The long and short of it is, I made a poor architectural decision when first refactoring the old ProjectShowcase component into the new ProjectBrowser component. I decided to keep the back button in the browser component and overlay it on the carousel. This kept the go back function from having to be passed as a prop to the carousel for callback, but it ended up causing me to add more complexity later that was simply not needed.
This was very evident when I wrote a callback function to check for the carousel component data state before rendering the back button. That led to needing to manage a separate state and start using hooks that were not really necessary.

So I fixed it. Moved the back button into the Carousel component where it belongs. Deleted all of that other crap. This is a good lesson for future me to be aware when refactoring and consider whether a shortcut in the moment might end up being a bad choice in the long run.

I have also removed some other code that was not adding any value: The useEffect that was resetting slide states. This was only ever needed in the old component before the refactor