Page Load updates
This mostly just adds a few reveal animations to some elements that were just being popped into view with no transition, but it also includes an experimental update to the camera controller component that starts the camera zoomed in and zooms out to the initial view as everything else is animating into view. I am not sure if I will end up keeping this change as I am on the fence about doing that much camera work on this project.