Brute force fix for the GLTF loader bug
For now, I am just setting the URIs for textures to the absolute path on the github deployment. This is definitely not my favorite solution, but it is faster than switching to the GLTF loader from three.js directly (as opposed to the drei hook I am currently using which does not support any URI configurations as far as I can tell)

The important thing to be aware of is that when I add a custom domain (shawnpavlas.work), the GLTF file will need to be changed again. Making a note of this in Notion docs.