Two small updates
Update 1: Moved Theme Switcher out of the Header component and into the MainContent component. This was because it really doesn't belong in the Header since it is not visually appearing in the header and also because we want to manage the visibility of it based directly on the isAnyCardOpen state of the MainContent, where it also happens to live much more comfortable.

Update 2: Set a default value for the limitMouseX utility function. This helps avoid errors by providing a fallback value and type inference.