backdrop-filter blur bug fix
After the most recent Chrome update, the backdrop filter blur that is used for the background of the maximized nav cards suddenly stopped working. I don't have any proof that the update caused it, but I know that reverting to earlier commits where I knew it was working previously seems to indicate this.
It looks like the cause of this is the default 0 opacity set on the nav-inner element which is a parent of the nav cards. This opacity 0 was set because I was animating in the section on page load, but because Chrome typically calculates everything at initial state and does not take into account the completed animation state it seems to have extended this behavior in the latest update to affect ancestors of elements. To clarify:
- Backdrop-filter blurs do not work on elements with opacity less than 0.
- Backdrop-filters are applied to the background of their element.
- In this latest update, it seems that because an ancestor of the element we wanted to use the blur effect on starts out at opacity 0 before animating into view, all descendants are unable to render blur properly.

Solution: Update animation keyframes and initial states to not rely on opacity. Worked around this by using transforms to move elements off screen for their initial state instead of having them start as invisible. This is a good workaround for these particular features