Quick ThemeProvider follow-up
My quick solution to adding a theme randomizer on page load had a particular flaw in that any new component that imported the context would start with a different random theme. This was because I was initializing the theme state with a calculated value rather than a constant one.
To address this, I created a new function that sets the theme to a random theme and can be accessed from context. This way, I can call the randomizer one time on the MainContent component and that will set the theme which can then be referenced correctly by any other components that use the theme context. This solution matches the pattern used for switching the theme and therefore makes most sense in the context of this project.